+++ src/main/java/com/github/jlangch/venice/impl/types/IVncJavaObject.java
@@ -26,6 +26,8 @@ import java.lang.reflect.Type;
 
 public interface IVncJavaObject {
 
+	boolean isEnum();
+
     Object getDelegate();
 
     default Class<?> getDelegateFormalType() { return null; }